% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometrySPoTApi.r
\name{FlowCytometrySPoTApi_add_atomic_flow_cytometry}
\alias{FlowCytometrySPoTApi_add_atomic_flow_cytometry}
\title{Create a single new object from a given data file with optional supplied metadata.}
\usage{
FlowCytometrySPoTApi_add_atomic_flow_cytometry(source, ...)
}
\description{
Create a single new object from a given data file with optional supplied metadata.
}
