% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometrySPoTApi.r
\name{FlowCytometrySPoTApi_add_flow_cytometry}
\alias{FlowCytometrySPoTApi_add_flow_cytometry}
\title{Create multiple new objects from a multi-row data file with optional supplied metadata}
\usage{
FlowCytometrySPoTApi_add_flow_cytometry(source, ...)
}
\arguments{
\item{source}{}
}
\description{
Create multiple new objects from a multi-row data file with optional supplied metadata
}
