% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometrySPoTApi.r
\name{FlowCytometrySPoTApi_delete_atomic}
\alias{FlowCytometrySPoTApi_delete_atomic}
\title{Delete the object}
\usage{
FlowCytometrySPoTApi_delete_atomic(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete the object
}
