% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometrySPoTApi.r
\name{FlowCytometrySPoTApi_get_group_curation_status}
\alias{FlowCytometrySPoTApi_get_group_curation_status}
\title{Get curation status of a group by ID (accession)}
\usage{
FlowCytometrySPoTApi_get_group_curation_status(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Get curation status of a group by ID (accession)
}
