% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCytometrySPoTApi.r
\name{FlowCytometrySPoTApi_search_runs}
\alias{FlowCytometrySPoTApi_search_runs}
\title{Retrieve run objects related to the given group}
\usage{
FlowCytometrySPoTApi_search_runs(id, page_offset, page_limit, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{page_offset}{Show the page {pageOffset+1} results from the start of the results. E.g. 100 will show a page of results  starting from the 101st result. The default value is 0.}

\item{page_limit}{Maximum number of results to return per page (see Paging above). This value must be between 0 and 2000 (inclusive). The default is 2000.}
}
\description{
Retrieve run objects related to the given group
}
